third_party/blink/web_tests/wpt_internal/attribution-reporting/register-source.https.html

<!doctype html>
<title>window.attributionReporting.registerSource</title>
<link rel="help" href="https://github.com/WICG/conversion-measurement-api/blob/main/event_attribution_reporting.md#registering-attribution-sources-with-javascript">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
'use strict';

promise_test(t => {
  return window.attributionReporting.registerSource('https://a.example');
}, 'valid params');

promise_test(t => {
 return promise_rejects_js(t, TypeError,
    window.attributionReporting.registerSource('xyz://a.example'));
}, 'invalid protocol');

promise_test(t => {
 return promise_rejects_js(t, TypeError,
    window.attributionReporting.registerSource('http://a.example'));
}, 'untrustworthy origin');

promise_test(t => {
  const iframe = document.createElement('iframe');
  document.body.appendChild(iframe);
  const frameWindow = iframe.contentWindow;
  const frameDOMException = frameWindow.DOMException;
  iframe.remove();

  return promise_rejects_dom(t, 'NotAllowedError', frameDOMException,
    frameWindow.attributionReporting.registerSource('https://a.example'));
}, 'window has no frame');

</script>
